% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dose_Combination_Code.R
\name{ApplyDrugComb}
\alias{ApplyDrugComb}
\title{\code{ApplyDrugComb} This function evaluates the p value under the different methos}
\usage{
ApplyDrugComb(levelsDrugA, levelsDrugB, Y, len, m)
}
\arguments{
\item{levelsDrugA}{no of dose groups for Drug A (r)}

\item{levelsDrugB}{no of dose groups for Drug B (s)}

\item{Y}{raw data}

\item{len}{Variable contain the number of patients in each dose combination group}

\item{m}{dose-response means}
}
\value{
returns the p-value vector
}
\description{
\code{ApplyDrugComb} This function evaluates the p value under the different methos
}
\references{
\url{http://stackoverflow.com/a/5577647}
}

