% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dose_Combination_Code.R
\name{DrugCombtest}
\alias{DrugCombtest}
\title{\code{DrugCombtest} drug combination tests are conducted here.}
\usage{
DrugCombtest(Y, len, levelsDrugB, levelsDrugA, ind, indcomb)
}
\arguments{
\item{Y}{obsservation for response}

\item{len}{Variable contain the number of patients in each dose combination group}

\item{levelsDrugB}{no of dose groups for Drug B (s)}

\item{levelsDrugA}{no of dose groups for Drug A (r)}

\item{ind}{all the drug dose combination groups (index including monotherapies and placebo control) in our study}

\item{indcomb}{all the drug dose combination groups (index excluding monotherapies and placebo control) in our study}
}
\value{
The object save in .
}
\description{
\code{DrugCombtest} drug combination tests are conducted here.
}
\references{
\url{http://stackoverflow.com/a/5577647}
}

