% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dose_Combination_Code.R
\name{eval_f_eq}
\alias{eval_f_eq}
\title{\code{eval_f_eq} This function evaluates the estimate under the null constraint (later used in function Est_eqrest)}
\usage{
eval_f_eq(x, C1, C2)
}
\arguments{
\item{x}{mean estimates under each dose group
The contraint c=MAX(a,b) is evaluated as (a+b-2c+|a-b|)=0}

\item{C1}{constraint matrix which evaluates a+b-2c}

\item{C2}{constraint matrix which evaluates a-b}
}
\value{
a vector
}
\description{
\code{eval_f_eq} This function evaluates the estimate under the null constraint (later used in function Est_eqrest)
}

